- Changed comment hard delete to soft delete: body and author are wiped - - Deleting a comment still does save information about who deleted it - - It also allowed for removal of one database query - Changed story rating displays to use cdn-img helper - Fixed story ratings selection in the story editor - Fixed file upload for a rating failure if rating had no image - Changed Json response serialization to return enums as their names and omit null data

+++ Ogma3/Api/V1/RatingsController.cs
@@ -80,7 +80,11 @@ namespace Ogma3.Api.V1
             
             if (rating.Icon != null && rating.Icon.Length > 0)
             {
-                await _b2Client.Files.Delete(r.IconId, r.Icon);
+                if (r.IconId != null && r.Icon != null)
+                {
+                    await _b2Client.Files.Delete(r.IconId, r.Icon);
+                }
+
                 var fileData = await _uploader.Upload(rating.Icon, "ratings", rating.Name+"_rating");
                 r.Icon = fileData.Path;
                 r.IconId = fileData.FileId;
